# backend/src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/user/create": {
      "post": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "   typeUser: 1 - Учитель   2 - Админ   3 - Ребенок   ",
            "schema": {
              "type": "object",
              "properties": {
                "tgUsername": {
                  "type": "string",
                  "example": "@string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "lastName": {
                  "type": "string",
                  "example": "string"
                },
                "secondName": {
                  "type": "string",
                  "example": "string"
                },
                "typeUser": {
                  "type": "number",
                  "example": 1
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "string"
                },
                "dateOfBirth": {
                  "type": "string",
                  "example": "2000-01-01"
                },
                "note": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/register": {
      "patch": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Вставка tg Id",
            "schema": {
              "type": "object",
              "properties": {
                "tgUsername": {
                  "type": "string",
                  "example": "@string"
                },
                "tgId": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{tgUsername}": {
      "get": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "tg Username"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Обновление данных",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "lastName": {
                  "type": "string",
                  "example": "string"
                },
                "secondName": {
                  "type": "string",
                  "example": "string"
                },
                "phoneNumber": {
                  "type": "string",
                  "example": "string"
                },
                "dateOfBirth": {
                  "type": "string",
                  "example": "2000-01-01"
                },
                "note": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/get/all": {
      "get": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "type",
            "description": "Какую группу пользователей нужно получить",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/get/role": {
      "get": {
        "tags": [
          "Пользователь"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "description": "Получить роль пользователя",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/create": {
      "post": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Создание учебной группы",
            "schema": {
              "type": "object",
              "properties": {
                "tgUsername": {
                  "type": "string",
                  "example": "@string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "urlName": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/get-all": {
      "get": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "in": "query",
            "description": "Фильтрация по Telegram username (можно передать несколько значений)",
            "type": "array",
            "collectionFormat": "multi",
            "items": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/add-childrens": {
      "post": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Создание учебной группы",
            "schema": {
              "type": "object",
              "properties": {
                "uuidGroup": {
                  "type": "string",
                  "example": "uuid"
                },
                "childrens": {
                  "type": "array",
                  "example": [
                    "tgUsername"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/{uuid}": {
      "delete": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/children/{tgUsername}/{uuidGroup}": {
      "delete": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "uuidGroup",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/add-teacher": {
      "post": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Назначение препода к группе",
            "schema": {
              "type": "object",
              "properties": {
                "tgUsername": {
                  "type": "string",
                  "example": "@string"
                },
                "uuid": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/update-group": {
      "put": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Обновление группы",
            "schema": {
              "type": "object",
              "properties": {
                "uuid": {
                  "type": "string",
                  "example": "string"
                },
                "name": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "string"
                },
                "urlName": {
                  "type": "string",
                  "example": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/education-group/get/composition/{uuid}": {
      "get": {
        "tags": [
          "Учебная группа"
        ],
        "description": "",
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/balance/update": {
      "patch": {
        "tags": [
          "Баланс"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Обновление баланса",
            "schema": {
              "type": "object",
              "properties": {
                "tgUsername": {
                  "type": "string",
                  "example": "@string"
                },
                "tgTeacher": {
                  "type": "string",
                  "example": "@string"
                },
                "operation": {
                  "type": "boolean",
                  "example": true
                },
                "summ": {
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/balance/get": {
      "get": {
        "tags": [
          "Баланс"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgUsername",
            "description": "По кому делать фильтрацию",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/balance/logs": {
      "get": {
        "tags": [
          "Баланс"
        ],
        "description": "",
        "parameters": [
          {
            "name": "tgTeacher",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          },
          {
            "name": "tgChild",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          },
          {
            "name": "operation",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          },
          {
            "name": "summ",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          },
          {
            "name": "startDate",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          },
          {
            "name": "equalSign",
            "description": "1 - равно 2 - операция меньше переданного числа 3 - операция больше переданного числа",
            "in": "query",
            "type": "string"
          },
          {
            "name": "endDate",
            "description": "Получение логов по фильтрам",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}